Refactor after review Move new enums to same nesting level as classes
