+++ Box.V2/Box.V2.csproj
@@ -54,6 +54,7 @@
     <Reference Include="System.IdentityModel.Tokens.Jwt, Version=6.12.2.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
       <HintPath>..\packages\System.IdentityModel.Tokens.Jwt.6.12.2\lib\net45\System.IdentityModel.Tokens.Jwt.dll</HintPath>
     </Reference>
+    <Reference Include="System.Runtime.Serialization" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -162,6 +163,7 @@
     <Compile Include="Models\BoxSignRequest.cs" />
     <Compile Include="Models\BoxSignRequestPrefillTag.cs" />
     <Compile Include="Models\BoxSignRequestSigner.cs" />
+    <Compile Include="Models\BoxSignRequestSignFiles.cs" />
     <Compile Include="Models\BoxStoragePolicy.cs" />
     <Compile Include="Models\BoxStoragePolicyAssignment.cs" />
     <Compile Include="Models\BoxTask.cs" />
