+++ Box.V2/Models/BoxSignRequest.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Runtime.Serialization;
 using Newtonsoft.Json;
 
 namespace Box.V2.Models
@@ -125,45 +126,25 @@ namespace Box.V2.Models
         /// </summary>
         [JsonProperty(PropertyName = FieldStatus)]
         public virtual BoxSignRequestStatus Status { get; private set; }
+    }
 
-        /// <summary>
-        /// Files that will be signed, which are copies of the original source files.
-        /// A new version of these files are created as signers sign and can be downloaded at any point in the signing process.
-        /// </summary>
-        public class BoxSignRequestSignFiles
-        {
-            public const string FieldFiles = "files";
-            public const string FieldIsReadyForDownload = "is_ready_for_download";
-
-            /// <summary>
-            /// Files that will be signed, which are copies of the original source files.
-            /// </summary>
-            [JsonProperty(PropertyName = FieldFiles)]
-            public virtual List<BoxFile> Files { get; private set; }
-
-            /// <summary>
-            /// Indicates whether the sign_files documents are processing and the PDFs may be out of date. A change to any document requires processing on all sign_files.
-            /// We recommended waiting until processing is finished (and this value is true) before downloading the PDFs.
-            /// </summary>
-            [JsonProperty(PropertyName = FieldIsReadyForDownload)]
-            public virtual bool IsReadyForDownload { get; private set; }
-        }
-
-        /// <summary>
-        /// Describes the status of the sign request.
-        /// </summary>
-        public enum BoxSignRequestStatus
-        {
-            converting,
-            created,
-            sent,
-            viewed,
-            signed,
-            cancelled,
-            declined,
-            error_converting,
-            error_sending,
-            expired
-        }
+    /// <summary>
+    /// Describes the status of the sign request.
+    /// </summary>
+    public enum BoxSignRequestStatus
+    {
+        converting,
+        created,
+        sent,
+        viewed,
+        signed,
+        cancelled,
+        declined,
+        error_converting,
+        error_sending,
+        expired,
+        downloaded,
+        [EnumMember(Value = "signed and downloaded")]
+        signed_and_downloaded
     }
 }
